chore(ui): rename index.css.tsx to styled.tsx
