+++ web/src/pages/BlockDetail/styled.tsx
